+++ CHANGELOG.md
@@ -63,6 +63,8 @@ This file follows the formats and conventions from [keepachangelog.com]
 
 * Experimental channel shape is now considered as a result of the configuration
   and not part of the measurement group configuration (#1296, #1466)
+* Use `LatestDeviceImpl` (currently `Device_5Impl`) for as a base class of the Sardana Tango
+  devices (#1214, #1301, #1531)
 
 ### Removed